Fixing the double format in the structure

+++ 0x00-math_complex/holberton.h
@@ -11,8 +11,8 @@
 typedef struct c
 {
 
-	int re;
-	int im;
+	double re;
+	double im;
 }
 complex;
